requirements.txt
tensorflow==1.15.0
tflearn==0.5.0
numpy==1.21.6
nltk==3.8.1
requests==2.31.0
pillow==9.5.0
typing-extensions==4.7.1
